Refactor ConfirmOverwrite
